Mobile sidemenu JS and CSS
Modified JS to check whether or not the desktop side menu is available (which at the mobile breakpoint it isn't), when it isn't it instead opens the mobile side menu. Then I styled the mobile side menu as close I could to the netmatters one. Currently white inner border is missing.