Awards Details fix
By taking the overflow property out of the owl carousel and putting it instead on the page border, it fixed the details not appearing. Also adjusted the dimensions of the details containers to make inside items not overflow. Adjusted the arrow psuedo element to position it in the middle. Added an invisible before element to extend the hitbox of the hover so it's easier to put cursor on the awards details.