Fixed Stickey Menu in Side menu
Now all of the stickey menu smoothly transitions left as expected. The problem was I wasn't manipulating the left property at the right place in the code. Also took away the smooth transition style while the menu is open so it doesn't keep occuring every time the user scrolls up.