PHP intergration
Started transferring project to PHP. Added the .env file, I don't know if it fully works yet. Created a database and linked it to the project. I'm not entirely sure if that fully works yet either. I also began transferring some of the reuseable code to includes. So far the header has been done.